#2388. Rename and reorder static_analysis_member_invocation_A06_t* tests
