+++ LanguageFeatures/Extension-types/static_analysis_member_invocation_A06_t03.dart
@@ -9,28 +9,65 @@
 ///   superinterface V that has an extension type member named n due to a member
 ///   declaration DM2, and DV does not declare a member that precludes DM2.
 ///
-/// @description Checks that if an extension type `ET` has a superinterface with
-/// a member `m` then this member is also present in `ET` but members of its
-/// representation type are not
+/// @description Checks that a getter doesn't preclude setter and vice versa
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=inline-class
 
-extension type ET0(int id) {
-  void m1() {}
-  int get m2 => 42;
-  void set m3(int val) {}
+import "../../Utils/expect.dart";
+
+String log = "";
+
+extension type V1(int _) {
+  String get n => "V1";
 }
 
-extension type ET1(int id) implements ET0 {}
+extension type V2(int _) {
+  void set n(String n) {
+    log = "V2: $n";
+  }
+}
+
+extension type V0(int _) implements V1, V2 {}
+
+extension type ET1(V1 _) implements V1 {
+  void set n(String v) {
+    log = "ET1: $v";
+  }
+}
+
+extension type ET2(V2 _) implements V2 {
+  String get n => "ET2";
+}
+
+extension type ET3(V0 _) implements V1, V2 {
+  String get n => "ET3";
+}
+
+extension type ET4(V0 _) implements V1, V2 {
+  void set n(String v) {
+    log = "ET4: $v";
+  }
+}
 
 main() {
-  ET1 et1 = ET1(42);
-  et1.m1();
-  et1.m2;
-  et1.m3 = 42;
-  et1.ceil();
-//    ^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
+  final v = V0(0);
+  Expect.equals("V1", ET1(v).n);
+  ET1(v).n = "a";
+  Expect.equals("ET1: a", log);
+  log = "";
+
+  Expect.equals("ET2", ET2(v).n);
+  ET2(v).n = "b";
+  Expect.equals("V2: b", log);
+  log = "";
+
+  Expect.equals("ET3", ET3(v).n);
+  ET3(v).n = "c";
+  Expect.equals("V2: c", log);
+  log = "";
+
+  Expect.equals("V1", ET4(v).n);
+  ET4(v).n = "d";
+  Expect.equals("ET4: d", log);
 }
