+++ LanguageFeatures/Extension-types/static_analysis_member_invocation_A06_t05.dart
@@ -9,7 +9,8 @@
 ///   superinterface V that has an extension type member named n due to a member
 ///   declaration DM2, and DV does not declare a member that precludes DM2.
 ///
-/// @description Checks that a getter doesn't preclude setter and vice versa
+/// @description Checks that an extension type declares a method or setter then
+/// they may preclude inherited members
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=inline-class
@@ -19,16 +20,14 @@ import "../../Utils/expect.dart";
 String log = "";
 
 extension type V1(int _) {
-  String get n => "V1";
+  String n() => "V1";
 }
 
 extension type V2(int _) {
-  void set n(String n) {
-    log = "V2: $n";
+  void set n(String v) {
+    log = "V2: $v";
   }
 }
-
-extension type V0(int _) implements V1, V2 {}
 
 extension type ET1(V1 _) implements V1 {
   void set n(String v) {
@@ -37,37 +36,25 @@ extension type ET1(V1 _) implements V1 {
 }
 
 extension type ET2(V2 _) implements V2 {
-  String get n => "ET2";
+  String n() => "ET2";
 }
 
-extension type ET3(V0 _) implements V1, V2 {
-  String get n => "ET3";
-}
-
-extension type ET4(V0 _) implements V1, V2 {
-  void set n(String v) {
-    log = "ET4: $v";
+extension type ET3(V1 _) implements V1 {
+  void set n(int v) {
+    log = "ET3: $v";
   }
 }
 
+extension type ET4(V2 _) implements V2 {
+  T n<T>(T t) => t;
+}
+
 main() {
-  final v = V0(0);
-  Expect.equals("V1", ET1(v).n);
-  ET1(v).n = "a";
-  Expect.equals("ET1: a", log);
+  ET1(V1(0)).n = "1";
+  Expect.equals("ET1: 1", log);
   log = "";
-
-  Expect.equals("ET2", ET2(v).n);
-  ET2(v).n = "b";
-  Expect.equals("V2: b", log);
-  log = "";
-
-  Expect.equals("ET3", ET3(v).n);
-  ET3(v).n = "c";
-  Expect.equals("V2: c", log);
-  log = "";
-
-  Expect.equals("V1", ET4(v).n);
-  ET4(v).n = "d";
-  Expect.equals("ET4: d", log);
+  Expect.equals("ET2", ET2(V2(0)).n());
+  ET3(V1(0)).n = 3;
+  Expect.equals("ET3: 3", log);
+  Expect.equals(42, ET4(V2(0)).n<int>(42));
 }
