+++ LanguageFeatures/Extension-types/static_analysis_member_invocation_A06_t06.dart
@@ -9,30 +9,23 @@
 ///   superinterface V that has an extension type member named n due to a member
 ///   declaration DM2, and DV does not declare a member that precludes DM2.
 ///
-/// @description Checks that an extension type declares a method or setter then
-/// they preclude inherited members
+/// @description Checks that if an extension type declares a method or setter
+/// and they preclude inherited members then the attempt to call of those
+/// precluded members is a compile-time error
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=inline-class
-
-import "../../Utils/expect.dart";
-
-String log = "";
 
 extension type V1(int _) {
   String n() => "V1";
 }
 
 extension type V2(int _) {
-  void set n(String v) {
-    log = "V2: $v";
-  }
+  void set n(String v) {}
 }
 
 extension type ET1(V1 _) implements V1 {
-  void set n(String v) {
-    log = "ET1: $v";
-  }
+  void set n(String v) {}
 }
 
 extension type ET2(V2 _) implements V2 {
@@ -40,21 +33,31 @@ extension type ET2(V2 _) implements V2 {
 }
 
 extension type ET3(V1 _) implements V1 {
-  void set n(int v) {
-    log = "ET3: $v";
-  }
+  void set n(int v) {}
 }
 
 extension type ET4(V2 _) implements V2 {
-  T n<T>(T t) => t;
+  int n() => 42;
 }
 
 main() {
-  ET1(V1(0)).n = "1";
-  Expect.equals("ET1: 1", log);
-  log = "";
-  Expect.equals("ET2", ET2(V2(0)).n());
-  ET3(V1(0)).n = 3;
-  Expect.equals("ET3: 3", log);
-  Expect.equals(42, ET4(V2(0)).n<int>(42));
+  ET1(V1(0)).n();
+//           ^
+// [analyzer] unspecified
+// [cfe] unspecified
+
+  ET2(V2(0)).n = "42";
+//           ^
+// [analyzer] unspecified
+// [cfe] unspecified
+
+  ET3(V1(0)).n();
+//           ^
+// [analyzer] unspecified
+// [cfe] unspecified
+
+  ET4(V2(0)).n = "42";
+//           ^
+// [analyzer] unspecified
+// [cfe] unspecified
 }
